requirements.txt
requests>=2.28.2
pandas>=1.5.3
numpy>=1.24.3
beautifulsoup4>=4.12.2
transformers>=4.30.2
torch>=2.0.1
python-dotenv>=1.0.0
setuptools>=65.0.0
feedparser>=6.0.10
